Fixing the issue mentioned in https://github.com/okta/okta-spring-boot/issues/25
The problem encountered was that the issuerUrl mentioned on the oktapreview.com admin domain no longer uses the /oauth2/default location.  The code was giving a StringIndexOutOfBoundsException on using the issuerUrl as mentioned on the UI.  I strongly feel there is no need for a base URL anymore and issuer URL is enough for the application to work.  The other workaround can be to check if the issuerUrl does contain /oauth2/, if it does extract the base url else continue with the issuerUrl.