Update to Spring Boot 2.5
With 2.5 a JwtDecoder bean is now initialized when `spring.security.oauth2.resourceserver.jwt.issuer-uri` is set.  There were a few invalid test assumptions that were setting this property (likely a copypasta eror)
For example the OAuth Code Flow ITs check to make sure requests to the /keys endpoint is NOT called (not needed for non OIDC or jwt resource server usage).

Another fix with a fragile test which detects the internal configuration of Spring Security.

OktaOAuth2Configurer had an unused else block, which caused a problem with 2.5 when a resource server was not configured.

Fixes: #280